fixed
Separated the functions responsible for handling contact requests into a separate controller file named contactsController.

Utilized controllers in the routes:
Replaced the functions within the routes with calls to the corresponding functions from the contactsController.

Applied authentication checks to all previously configured routes:
Added the authMiddleware to all routes that require authentication to ensure the security of user data.